requirements.txt
sqlalchemy
python-telegram-bot==5.3.0
requests==2.12.4
psycopg2
telepot